# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /skills:
    get:
      tags: [Skills]
      operationId: getSkills
      summary: Get skills configuration for the Skills page
      responses:
        '200':
          description: Skills configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillsConfig'
    put:
      tags: [Skills]
      operationId: upsertSkills
      summary: Create or update the skills configuration (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillsConfig'
      responses:
        '200':
          description: Upserted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillsConfig'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /sections/{key}:
    get:
      tags: [Sections]
      operationId: getSection
      summary: Get a single section by key
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The section
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Section'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    put:
      tags: [Sections]
      operationId: upsertSection
      summary: Create or update a section by key (admin)
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertSectionInput'
      responses:
        '200':
          description: Upserted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Section'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Section:
      type: object
      required: [key, value, updatedAt]
      properties:
        key:
          type: string
        value:
          type: string
        updatedAt:
          type: string
          format: date-time
    UpsertSectionInput:
      type: object
      required: [value]
      properties:
        value:
          type: string
    Skill:
      type: object
      required: [id, name, level]
      properties:
        id:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        level:
          type: integer
          minimum: 0
          maximum: 100
    SkillCategory:
      type: object
      required: [id, title, icon, skills]
      properties:
        id:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        icon:
          type: string
          minLength: 1
          description: Lucide icon name used by the frontend (e.g. Cpu, CircuitBoard).
        skills:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
    SkillsConfig:
      type: object
      required: [categories]
      properties:
        categories:
          type: array
          items:
            $ref: '#/components/schemas/SkillCategory'
...