# lib/api-zod/src/generated/types/skill.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Skill {
  /** @minLength 1 */
  id: string;
  /** @minLength 1 */
  name: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  level: number;
}